+++ test/dummy/config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+  mount RecordingStudioRootSwitchable::Engine, at: "/recording_studio_root_switchable"
   devise_for :users
   resources :access_actions, only: [ :show ], param: :action_name
   resources :message_groups, only: [ :index ]